# api/docs/swagger.yaml
basePath: /api
definitions:
  handlers.CreateDeviceBody:
    properties:
      device_code:
        example: esp32-001
        type: string
      name:
        type: string
      vehicle_name:
        type: string
      vehicle_plate:
        type: string
    type: object
  models.Alert:
    properties:
      alert_type:
        type: string
      id:
        type: string
      message:
        type: string
      occurred_at:
        type: string
      resolved_at:
        type: string
      severity:
        type: string
      threshold:
        type: number
      value:
        type: number
    type: object
  models.Device:
    properties:
      created_at:
        type: string
      device_code:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
      vehicle_name:
        type: string
      vehicle_plate:
        type: string
    type: object
  models.TelemetryRow:
    properties:
      battery_voltage:
        type: number
      coolant_temp:
        type: number
      dtc_count:
        type: integer
      engine_load:
        type: number
      id:
        type: integer
      latitude:
        type: number
      longitude:
        type: number
      mil_status:
        type: boolean
      recorded_at:
        type: string
      rpm:
        type: integer
      speed:
        type: number
      throttle:
        type: number
    type: object
  models.Trip:
    properties:
      avg_speed:
        type: number
      duration_seconds:
        type: integer
      ended_at:
        type: string
      id:
        type: string
      max_speed:
        type: number
      started_at:
        type: string
    type: object
info:
  contact: {}
  description: REST API for OBD2 devices, telemetry, trips, and alerts.
  title: ESP32 OBD2 Cloud API
  version: "1.0"
paths:
  /devices:
    get:
      description: Returns all registered OBD2 devices ordered by device code.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Device'
            type: array
        "500":
          description: error message
          schema:
            type: string
      summary: List devices
      tags:
      - devices
    post:
      consumes:
      - application/json
      description: Creates a device row; device_code must be unique and match MQTT
        topic segments.
      parameters:
      - description: Device payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateDeviceBody'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Device'
        "400":
          description: invalid json or validation error
          schema:
            type: string
      summary: Create device
      tags:
      - devices
  /devices/{id}:
    get:
      description: Returns device metadata for the given id.
      parameters:
      - description: Device UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Device'
        "400":
          description: invalid id
          schema:
            type: string
        "404":
          description: not found
          schema:
            type: string
        "500":
          description: error message
          schema:
            type: string
      summary: Get device
      tags:
      - devices
  /devices/{id}/alerts:
    get:
      description: Returns up to `limit` alerts (default 50, max 200), newest first.
      parameters:
      - description: Device UUID
        in: path
        name: id
        required: true
        type: string
      - description: Max rows (default 50, max 200)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Alert'
            type: array
        "400":
          description: invalid id
          schema:
            type: string
        "404":
          description: device not found
          schema:
            type: string
        "500":
          description: error message
          schema:
            type: string
      summary: List alerts
      tags:
      - alerts
  /devices/{id}/latest:
    get:
      description: Returns the newest telemetry row for the device.
      parameters:
      - description: Device UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TelemetryRow'
        "400":
          description: invalid id
          schema:
            type: string
        "404":
          description: device or telemetry not found
          schema:
            type: string
        "500":
          description: error message
          schema:
            type: string
      summary: Latest telemetry
      tags:
      - telemetry
  /devices/{id}/telemetry:
    get:
      description: Returns up to `limit` rows (default 100, max 500).
      parameters:
      - description: Device UUID
        in: path
        name: id
        required: true
        type: string
      - description: Max rows (default 100, max 500)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.TelemetryRow'
            type: array
        "400":
          description: invalid id
          schema:
            type: string
        "404":
          description: device not found
          schema:
            type: string
        "500":
          description: error message
          schema:
            type: string
      summary: List telemetry
      tags:
      - telemetry
  /devices/{id}/trips:
    get:
      description: Returns up to `limit` trips (default 50, max 200), newest first.
      parameters:
      - description: Device UUID
        in: path
        name: id
        required: true
        type: string
      - description: Max rows (default 50, max 200)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Trip'
            type: array
        "400":
          description: invalid id
          schema:
            type: string
        "404":
          description: device not found
          schema:
            type: string
        "500":
          description: error message
          schema:
            type: string
      summary: List trips
      tags:
      - trips
swagger: "2.0"